+++ widgets/index.ts
@@ -2,3 +2,4 @@ export * from "./Frame";
 export * from "./InfoBox";
 export * from "./ScrollablePanel";
 export * from "./TabbedScrollablePanel";
+export * from "./Tree";